cleanup

+++ TradeHero/TH_ANDROID/src/main/java/com/tradehero/common/widget/ColorIndicator.java
@@ -13,6 +13,7 @@ public class ColorIndicator extends RelativeLayout
 {
     protected static final int PERCENT_STRETCHING_FOR_COLOR = 20;
 
+    //<editor-fold desc="Constructors">
     public ColorIndicator(Context context)
     {
         super(context);
@@ -27,14 +28,18 @@ public class ColorIndicator extends RelativeLayout
     {
         super(context, attrs, defStyle);
     }
+    //</editor-fold>
 
     public void linkWith(Double percentage)
     {
-        if (percentage == null || percentage.doubleValue() == 0.0) {
+        if (percentage == null || percentage == 0.0)
+        {
             setBackgroundColor(getContext().getResources().getColor(R.color.gray_2));
-        } else {
+        }
+        else
+        {
             setBackgroundColor(
-                    ColorUtils.getColorForPercentage((float)percentage.doubleValue() * PERCENT_STRETCHING_FOR_COLOR));
+                    ColorUtils.getColorForPercentage((float) percentage.doubleValue() * PERCENT_STRETCHING_FOR_COLOR));
         }
     }
 }
